# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Binance Learning Agent Source API",
    "version": "1.0.0",
    "description": "API para que ChatGPT consulte el source de aprendizaje de arbitraje educativo en Binance y pueda crear progreso, notas, borradores, cursos, modulos y clases."
  },
  "servers": [
    {
      "url": "https://binancelearning-production.up.railway.app",
      "description": "Railway production"
    }
  ],
  "security": [
    {
      "AgentApiKey": []
    }
  ],
  "components": {
    "securitySchemes": {
      "AgentApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-agent-api-key",
        "description": "API key configured in AGENT_API_KEY."
      }
    },
    "schemas": {
...
      "ModuleInput": {
        "type": "object",
        "required": ["title", "summary"],
        "properties": {
          "courseId": { "type": "string" },
          "courseSlug": { "type": "string" },
          "title": { "type": "string" },
          "summary": { "type": "string" },
          "sortOrder": { "type": "integer", "default": 0 }
        },
        "description": "Provide either courseId or courseSlug."
      },
      "LessonInput": {
        "type": "object",
        "required": ["moduleId", "title", "objective", "content"],
        "properties": {
          "moduleId": { "type": "string" },
          "title": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["THEORY", "PRACTICE"],
            "default": "THEORY"
          },
          "difficulty": {
            "type": "string",
            "enum": ["BEGINNER", "INTERMEDIATE", "ADVANCED"],
            "default": "BEGINNER"
          },
          "durationMinutes": { "type": "integer", "minimum": 5, "maximum": 240, "default": 20 },
          "objective": { "type": "string" },
          "content": { "type": "string" },
          "practicePrompt": { "type": "string", "nullable": true },
          "successCriteria": { "type": "string", "nullable": true },
          "resources": {
            "type": "array",
            "items": { "type": "string" }
          },
          "sortOrder": { "type": "integer", "default": 0 }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/agent/modules": {
      "post": {
        "operationId": "createModule",
        "summary": "Create a course module",
        "description": "Allows ChatGPT to add a new module to a course.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ModuleInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Module created" },
          "404": { "description": "Course not found" }
        }
      }
    },
    "/api/agent/lessons": {
      "post": {
        "operationId": "createLesson",
        "summary": "Create a theory or practice lesson",
        "description": "Allows ChatGPT to write new lesson content under an existing module.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LessonInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Lesson created" },
          "404": { "description": "Module not found" }
        }
      }
    },
...
  }
}